LPD-58182 Merge GroupByTranslator/DefaultGroupByTranslator/GroupByTranslatorImpl into pojos
